+++ tests/Command/AddLinkCommandTest.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SocialFoodSolutions\Test\Command;
+
+use PHPUnit\Framework\TestCase;
+
+class AddLinkCommandTest extends TestCase
+{
+    public function ()
+    {
+        
+    }
+}